+++ src/toolbar/mark-button.tsx
@@ -1,38 +1,37 @@
-import { type FunctionComponent, type PointerEvent, useCallback } from "react";
-import { Button, Tooltip } from "@shopify/polaris";
+import { type PointerEvent, useCallback, useId } from "react";
 import { useFocused, useSlate } from "slate-react";
 import { type CustomTextKey } from "~/types";
 import { isMarkActive, toggleMark } from "~/helper/mark";
 
 type MarkButtonProps = {
   format: CustomTextKey;
-  icon: FunctionComponent;
 }
 
-export function MarkButton({ format, icon }: MarkButtonProps) {
+export function MarkButton({ format }: MarkButtonProps) {
   const editor = useSlate();
   const focused = useFocused();
   const isActive = isMarkActive(editor, format) && focused;
+  const tooltipId = useId();
 
-  const handlePointerDown = useCallback((event: PointerEvent<HTMLSpanElement>) => {
+  const handleOnClick = useCallback((event: PointerEvent<HTMLSpanElement>) => {
     event.preventDefault();
     toggleMark(editor, format);
   }, [editor, format]);
 
-  const tooltipMap = {
-    bold: 'Bold',
-    italic: 'Italic'
-  }
-
   return (
-    <Tooltip content={tooltipMap[format]}>
-      <Button
-        variant="plain"
-        accessibilityLabel={isActive ? `Remove ${format}` : `Add ${format}`}
-        pressed={isActive}
-        onPointerDown={handlePointerDown}
-        icon={icon}
-      />
-    </Tooltip>
+    <>
+      <s-button 
+        onClick={handleOnClick}
+        variant="tertiary"
+        interestFor={tooltipId}
+        icon={ format === 'bold' ? 'text-bold' : 'text-italic' }
+        accessibilityLabel={isActive ? 'Remove formatting' : 'Add formatting'}
+      >
+      </s-button>
+
+      <s-tooltip id={tooltipId}>
+        <s-text>{ format === 'bold' ? 'Bold' : 'Italic' }</s-text>
+      </s-tooltip>
+    </>
   )
 }